requirements.txt
gradio>=4.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
tensorflow>=2.12.0
numpy>=1.22.0
pillow>=9.0.0
python-multipart>=0.0.6
mediapipe>=0.10.0
opencv-python-headless>=4.5.0